+++ frontend/src/types/enums.ts
@@ -1,0 +1,4 @@
+export enum HeadingEnum {
+  section = "section",
+  subSection = "subSection",
+}
