+++ frontend/src/types/types.ts
@@ -3,7 +3,3 @@ import { guestSchema } from "./schemas";
 
 export type GuestType = z.infer<typeof guestSchema>;
 
-export enum HeadingEnum {
-  section = "section",
-  subSection = "subSection",
-}
